# data-platform-observability-service/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Data Platform Observability Service",
    "description": "Phase 9 operational control plane: observes, correlates, and evaluates SLA/health/alerts across Phases 1-8's independently-owned workflows. Owns no sibling's workflow execution — /v1/customer/* is customer-safe; /internal/v1/* and /v1/events are platform/service-to-service only.",
    "version": "0.1.0"
  },
  "components": {
    "securitySchemes": {
      "internalApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-internal-api-key"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
  },
  "paths": {
...
    "/internal/v1/operations/data-products/{productId}/executions": {
      "get": {
        "tags": [
          "internal",
          "operations"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "productId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/internal/v1/operations/executions/{executionId}": {
      "get": {
        "tags": [
          "internal",
          "operations"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "executionId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/internal/v1/operations/executions/{executionId}/timeline": {
      "get": {
        "tags": [
          "internal",
          "operations"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "executionId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
    "/internal/v1/operations/alerts/{alertId}/ack": {
      "post": {
        "tags": [
          "internal",
          "alerts"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "alertId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/internal/v1/operations/alerts/{alertId}/resolve": {
      "post": {
        "tags": [
          "internal",
          "alerts"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "alertId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/internal/v1/operations/alerts/{alertId}/suppress": {
      "post": {
        "tags": [
          "internal",
          "alerts"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "alertId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
  }
}